feat: move codecov-cli to uv to simplify dependency management and build APIs (#28)
I did try out uv workspaces to manage both with uv commands from the
project root, but I found it a bit blah without totally changing the
project structure. We can revisit that later if we want. For now the
output of this is moving codecov-cli to uv for dependency management and
build.

README overhaul to come later with documentation of the commands here.
Just trying to get our builds published first.

Next PR will update the build and publish workflows.