fix: codecov-cli build flow (#29)
Fixes codecov-cli build flow. It's possible something goes wrong with
the publish steps themselves as I haven't tested those, but the builds
should all be working.

Note that with the removal of the treesitter deps, we're now building a
pure python wheel and so only need one bdist (`py3-none-any`). The
executable builds are largely unaffected, just updating to work with uv
for the most part.