feat: Add codecov-cli gh release publish to craft flow (#83)
Doing these separately caused issues as craft was deleting the
branch/references to the codecov cli assets. This pr makes the
codecov-cli assets upload with the prevent assets through the craft
flow.

Also adds a script to change the editable field on the codecov-cli
dependency in prevent-cli, which is necessary for the pyinstaller build
to work.