build: pin click version to correct version (#95)
I manually tested that 8.3.0 is causing the commit sha not found cli
issue so i'm manually pinning the version to avoid that for now. not
sure what the root cause is, just treating the symptom that 8.3 breaks
things for now. See previous commits on this PRs' CI runs for the proof
that 8.3 is broken